+++ mycodo/actions/create_note.py
@@ -53,6 +53,13 @@ ACTION_INFORMATION = {
             'required': False,
             'name': lazy_gettext('Note'),
             'phrase': 'The body of the note'
+        },
+        {
+            'id': 'include_message',
+            'type': 'bool',
+            'default_value': False,
+            'name': 'Include Message in Note',
+            'phrase': "Include the message passed to the action in the note that's created"
         }
     ]
 }
@@ -66,6 +73,7 @@ class ActionModule(AbstractFunctionAction):
         self.tag = None
         self.name = None
         self.note = None
+        self.include_message = None
 
         action = db_retrieve_table_daemon(
             Actions, unique_id=self.unique_id)
@@ -96,44 +104,37 @@ class ActionModule(AbstractFunctionAction):
         except:
             note = self.note
 
+        if self.include_message and 'message' in dict_vars:
+            note += f"\n\n{dict_vars['message']}"
+
         self.logger.debug(f"Tag(s): {','.join(list_tags)}, name: {name}, note: '{note}'.")
 
         with session_scope(MYCODO_DB_PATH) as new_session:
-            list_tag_names = []
+            list_tag_ids = []
             for each_tag in list_tags:
-                # First check name
                 tag_check = new_session.query(NoteTags).filter(
-                    NoteTags.name == each_tag).first()
+                    NoteTags.unique_id == each_tag).first()
                 if tag_check:
-                    list_tag_names.append(tag_check.name)
+                    list_tag_ids.append(each_tag)
                 else:
-                    # Next check ID
-                    tag_check = new_session.query(NoteTags).filter(
-                        NoteTags.unique_id == each_tag).first()
-                    if tag_check:
-                        list_tag_names.append(tag_check.name)
-                    else:
-                        self.logger.error(
-                            f"Tag with name or id '{each_tag}' does not exist.")
+                    self.logger.error(f"Tag with ID '{each_tag}' does not exist.")
 
-            if not list_tag_names:
+            if not list_tag_ids:
                 msg = "No valid tags specified. Cannot create note."
                 dict_vars['message'] += msg
                 self.logger.error(msg)
                 return dict_vars
 
-            dict_vars['message'] += f" Create note with name '{name}', tag(s) '{','.join(list_tag_names)}'"
+            dict_vars['message'] += f" Create note with name '{name}', tag ID(s) '{','.join(list_tag_ids)}'"
             if note:
                 dict_vars['message'] += f", and note {note}"
             dict_vars['message'] += "."
 
             new_note = Notes()
             new_note.name = name
-            new_note.tags = ','.join(list_tag_names)
+            new_note.tags = ','.join(list_tag_ids)
             if note:
                 new_note.note = note
-            else:
-                new_note.note = dict_vars['message']
             new_session.add(new_note)
 
         self.logger.debug(f"Message: {dict_vars['message']}")